--- Signal/src/util/OWSAsserts.swift
@@ -1,9 +1,0 @@
-//
-//  Copyright (c) 2017 Open Whisper Systems. All rights reserved.
-//
-
-func owsFail(_ message: String) {
-    Logger.error(message)
-    Logger.flush()
-    assertionFailure(message)
-}
